backend/requirement.txt
blinker==1.9.0
click==8.1.8
dnspython==1.16.0
Flask==3.1.0
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.11.2
python-dotenv==1.0.1
Werkzeug==3.1.3
